refactor(kernel-node): share the temp-write ceremony behind the lock's two atomic publishes
- writeTempThenPublish holds the temp naming, 0o600 mode, fsync, and cleanup once; the claim (link) and the port file (rename) inject only their publish verb
- contentionAgainstBound rewritten async/await; drops the inert | never union
