+++ packages/kernel-node/src/lock/port-file.ts
@@ -1,35 +1,11 @@
-import { open, rename, unlink } from "node:fs/promises";
+import { rename } from "node:fs/promises";
 import { readFileSync } from "node:fs";
-import { randomBytes } from "node:crypto";
+import { writeTempThenPublish } from "./write-temp.js";
 
 export async function writePortFile(portFilePath: string, port: number): Promise<void> {
-  const serialized = `${port}\n`;
-  const suffix = randomBytes(4).toString("hex");
-  const tempPath = `${portFilePath}.tmp.${suffix}`;
-
-  let fh: Awaited<ReturnType<typeof open>> | null = null;
-  try {
-    fh = await open(tempPath, "w", 0o600);
-    await fh.writeFile(serialized, "utf-8");
-    await fh.sync();
-    await fh.close();
-    fh = null;
-    await rename(tempPath, portFilePath);
-  } catch (err) {
-    if (fh !== null) {
-      try {
-        await fh.close();
-      } catch {
-        /* already in the error path */
-      }
-    }
-    try {
-      await unlink(tempPath);
-    } catch {
-      /* temp sibling may not exist */
-    }
-    throw err;
-  }
+  await writeTempThenPublish(portFilePath, `${port}\n`, (tempPath, targetPath) =>
+    rename(tempPath, targetPath),
+  );
 }
 
 export function readPortFile(portFilePath: string): number | null {
